backend/requirements.txt
fastapi
uvicorn
langgraph
langchain-google-genai
sqlalchemy
pydantic
pydantic-settings
python-dotenv
